requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
matplotlib==3.8.0
python-dotenv==1.0.0
pydantic==2.4.2
numba==0.59.1